fix bug in job script file

+++ batch/runbatch_spikeSorting.sh
@@ -36,7 +36,10 @@ cd /u/home/x/xinniu/nwbPipeline/batch
 
 # make a copy of batch script so that we can submit jobs for other patients
 # when previous jobs are still running:
-cp runbatch_spikeSorting.m runbatch_spikeSorting_$JOB_ID.m
+
+if [ ! -f "$file" ]; then
+    cp runbatch_spikeSorting.m runbatch_spikeSorting_$JOB_ID.m
+fi
 
 matlab  -nosplash -nodisplay -singleCompThread <<EOF
     runbatch_spikeSorting_$JOB_ID($SGE_TASK_ID, $total_tasks);
